CM-1631: Implemented null-safe conversion from Long to Instant and Duration. DataUtils renamed into DataModelUtils for better clarity.
