+++ comet-java-client/src/main/java/ml/comet/experiment/impl/BaseExperimentAsync.java
@@ -43,16 +43,16 @@ import static ml.comet.experiment.impl.resources.LogMessages.LOG_REMOTE_ASSET_UR
 import static ml.comet.experiment.impl.resources.LogMessages.getString;
 import static ml.comet.experiment.impl.utils.AssetUtils.createAssetFromData;
 import static ml.comet.experiment.impl.utils.AssetUtils.createAssetFromFile;
-import static ml.comet.experiment.impl.utils.DataUtils.createGitMetadataRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createGraphRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogEndTimeRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogHtmlRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogLineRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogMetricRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogOtherRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogParamRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createLogStartTimeRequest;
-import static ml.comet.experiment.impl.utils.DataUtils.createTagRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createGitMetadataRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createGraphRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogEndTimeRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogHtmlRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogLineRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogMetricRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogOtherRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogParamRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createLogStartTimeRequest;
+import static ml.comet.experiment.impl.utils.DataModelUtils.createTagRequest;
 
 /**
  * The base class for all asynchronous experiment implementations providing implementation of common routines
