+++ comet-java-client/src/main/java/ml/comet/experiment/impl/utils/CometUtils.java
@@ -6,6 +6,9 @@ import ml.comet.experiment.context.ExperimentContext;
 import org.apache.commons.lang3.StringUtils;
 
 import java.net.URI;
+import java.time.Duration;
+import java.time.Instant;
+import java.time.temporal.ChronoUnit;
 import java.util.Map;
 import java.util.UUID;
 
@@ -86,4 +89,31 @@ public final class CometUtils {
             return String.format("%s_%s", context.getContext(), name);
         }
     }
+
+    /**
+     * Allows safe conversion of timestamp into {@link Instant}.
+     *
+     * @param timestamp the time in millis since epoch or {@code null}.
+     * @return the instance of {@link Instant} or {@code null} if timestamp is not set.
+     */
+    public static Instant instantOrNull(Long timestamp) {
+        if (timestamp != null) {
+            return Instant.ofEpochMilli(timestamp);
+        }
+        return null;
+    }
+
+    /**
+     * Allows safe conversion of duration in millis into {@link Duration}.
+     *
+     * @param durationMillis the duration in millis since epoch or {@code null}.
+     * @return the instance of {@link Duration} or {@code null} if duration is not set.
+     */
+    public static Duration durationOrNull(Long durationMillis) {
+        if (durationMillis != null) {
+            return Duration.of(durationMillis, ChronoUnit.MILLIS);
+        }
+        return null;
+    }
+
 }
